more dtype issues

+++ src/transformers/models/hunyuan_v1_moe/modeling_hunyuan_v1_moe.py
@@ -291,10 +291,10 @@ class HunYuanMoEV1Moe(nn.Module):
         self.shared_mlp = HunYuanMoEV1MLP(config)
 
     def route_tokens_to_experts(self, hidden_states):
-        routing_weights = F.softmax(hidden_states, dim=1)
+        routing_weights = F.softmax(hidden_states, dim=1, dtype=torch.float)
         routing_weights, selected_experts = torch.topk(routing_weights, self.top_k, dim=-1)
         routing_weights /= routing_weights.sum(dim=-1, keepdim=True)
-        return selected_experts, routing_weights
+        return selected_experts, routing_weights.to(hidden_states.dtype)
 
     def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
         batch_size, sequence_length, hidden_dim = hidden_states.shape
